add the missed logic to the order

+++ Catalogs.Domain/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace Catalogs.Domain.Entities
         public Guid Id { get; set; }
         public Guid BaseItemId { get; set; }
         public string SKU { get; set; } = null!;
-        public int StockQuantity { get; set; }
+        public double StockQuantity { get; set; }
         public decimal? DiscountPrice { get; set; }
         public bool HasDiscount => DiscountPrice.HasValue;
         public List<ProductFeature> Features { get; set; } = new();
