+++ Shared.Contracts/DTOs/ProductDetailsDTO.cs
@@ -7,7 +7,7 @@
         public double Price { get; set; }
         public Guid CategoryId { get; set; }
         public string SKU { get; set; }
-        public int StockQuantity { get; set; }
+        public double StockQuantity { get; set; }
         public bool IsAvailable { get; set; }
         public Guid UserId { get; set; }
         public DateTime CreatedAt { get; set; }
